behindScheduleCapacity is being calculated und Output property workstationsWithOvertime was renamed to workstations

+++ src/main/java/com/ibsys/backend/core/domain/entity/Workplace.java
@@ -36,7 +36,7 @@ public class Workplace {
     private String machineidletimecosts;
 
     @Column(name = "timeneed")
-    private long timeneed;
+    private int timeneed;
 
     @Column(name = "period")
     private int period;
